+++ index.js
@@ -1,12 +1,9 @@
-const express = require('express');
-const bodyParser = require('body-parser');
-const path = require('path');
 const mongoose = require('mongoose');
 const keys = require('./config/keys');
 const WebSocket = require('ws');
 const http = require('http');
 
-const app = express();
+const { app } = require('./app');
 const Stock = require('./models/Stocks');
 
 //Server is a separate http server to bind app to
@@ -19,10 +16,7 @@ mongoose.connect(keys.mongoURI, {
   useUnifiedTopology: true,
 });
 
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-
-//Server to Client Connection
+//Websocket Server to Client Connection
 wsServer.on('connection', ws => {
   ws.send(JSON.stringify({ type: 'message', data: 'Hello Client' }));
   console.log('Client connected');
@@ -92,19 +86,6 @@ wsServer.on('connection', ws => {
     ws.terminate();
   });
 });
-
-if (process.env.NODE_ENV === 'production') {
-  //Express will serve up production assets
-  //like our main.js file, or main.css file
-  app.use(express.static('client/build'));
-
-  // Express will serve up the index.html file
-  // if it doesn't recognize the route
-
-  app.get('*', (req, res) => {
-    res.sendFile(path.resolve(__dirname, 'client', 'build', 'index.html'));
-  });
-}
 
 const PORT = process.env.PORT || 5000;
